TVM CLI to compile models and generate config file
Python script to compile a model using TVM and generate a C++ header
file to be used in the inference engine. Support .onnx/.pb models.
The tool takes the model zoo definition.yaml file as an input

A dockerfile is provided to build an environment with all required
dependencies to run the cli tool and build the example pipeline for
yolo_v2_tiny. The upstream TVM docker images are not maintained very
well; hence we manually build and install tvm to be able to use a newer
version and more reproducible environment.

Change-Id: Ie74f9aef4ebacd8a9f446ea88f6e8443be6238ec
Signed-off-by: Luca Foschiani <luca.foschiani@arm.com>
Issue-Id: SCM-874